LPS-91923 Move EDITABLE_FRAGMENT_ENTRY_PROCESSOR to constants
